+++ fileware/server.py
@@ -1,16 +1,13 @@
 import base64
 import binascii
 import os
-from datetime import datetime
 from http.server import SimpleHTTPRequestHandler
 from pathlib import PurePath
 from time import time
-from urllib.request import urlopen
 
 import yaml
-from gmailconnector.send_email import SendEmail
 
-from . import env, models, ngrok
+from . import env, models, settings
 
 logger = models.server_logger()
 
@@ -57,7 +54,7 @@ class Authenticator(SimpleHTTPRequestHandler):
         super().__init__(*args, **kwargs)
 
     def log_message(self, format_: str, *args: tuple) -> None:
-        """Suppresses logs from http.server by holding the args.
+        """Suppresses logs from ``http.server`` by holding the args.
 
         Overwrites the base method: `BaseHTTPRequestHandler.log_message()`
 
@@ -65,6 +62,8 @@ class Authenticator(SimpleHTTPRequestHandler):
             format_: String operator %s
             *args: Logs from SimpleHTTPRequestHandler displaying request method type and HTTP status code.
         """
+        if not hasattr(self, 'path'):
+            return
         if len(args) == 3:  # We don't have to care about NOT FOUND items because they are mostly hidden/protected.
             method, status_code, ignore = args  # ignore always returns `-`
             method = str(method).split('/')[0].strip()
@@ -79,6 +78,7 @@ class Authenticator(SimpleHTTPRequestHandler):
         self.send_header("Content-type", "text/html")
         self.end_headers()
 
+    # noinspection PyProtectedMember
     def do_GET(self) -> None:
         """Serve a front end with user authentication."""
         if _reset_auth() and 'Authorization' in self.headers.keys():
@@ -88,7 +88,7 @@ class Authenticator(SimpleHTTPRequestHandler):
 
         if not (auth_header := self.headers.get("Authorization")):
             self.do_AUTH()
-            if not env.authenticated:
+            if not settings._authenticated:
                 self.wfile.write(login_failed.encode(encoding='UTF-8'))
             else:
                 self.wfile.write(session_expiry.encode(encoding='UTF-8'))
@@ -99,16 +99,16 @@ class Authenticator(SimpleHTTPRequestHandler):
                 new_name = target_path + 'index_TEMP.html'
                 logger.critical(f'Renaming {old_name} to {new_name}')
                 os.rename(old_name, new_name)
-                env.renamed.append({old_name: new_name})
-            if not env.authenticated:
+                settings._renamed.append({old_name: new_name})
+            if not settings._authenticated:
                 self.do_HEAD()
                 self.wfile.write(auth_success.encode(encoding='UTF-8'))
-                env.authenticated = True
+                settings._authenticated = True
                 return
             try:
                 SimpleHTTPRequestHandler.do_GET(self)
             except BrokenPipeError:
-                logger.error(f'Received BrokenPipeError while reaching {self.path}')
+                pass
         else:
             self.do_AUTH()
             auth = auth_header.strip('Basic ')
@@ -127,71 +127,20 @@ class Authenticator(SimpleHTTPRequestHandler):
 
     def do_POST(self) -> None:
         """Handles POST request and writes the received data into a yaml file."""
-        now = datetime.now()
-        client_file = 'client_info.yaml'
         length = int(self.headers.get('content-length'))
         peanut = self.rfile.read(length)
         butter = peanut.decode(encoding='UTF-8').split('&')[0]
         if butter == 'LOGOUT':
             logger.critical("Logout has been requested. Resetting auth headers to False.")
-            env.first_run = True  # to reset and give a fresh start
-            env.authenticated = False  # to show Login Success screen the next time
+            settings._first_run = True  # to reset and give a fresh start
+            settings._authenticated = False  # to show Login Success screen the next time
             return
         client_info = yaml.load(butter, Loader=yaml.FullLoader)
-        if client_info.get('ip') == yaml.load(urlopen('https://ipapi.co/json/'), Loader=yaml.FullLoader).get('ip'):
-            logger.info(f"Internal connection request received. Response: {client_info.get('_html_ref')}")
+        filename = f"{client_info.get('ip')}.yaml"
+        if os.path.isfile(filename):
             return
-        logger.fatal(str(client_info).strip("{}").replace("'", ""))
-
-        if os.path.isfile(client_file):
-            with open(client_file, 'r') as client:
-                exist = yaml.load(''.join([nut for nut in client.readlines() if not nut.startswith('#')]),
-                                  Loader=yaml.FullLoader)
-
-            # avoids duplicate notification for repeated trials in less than 5 minutes
-            if int(now.timestamp()) - int(os.stat(client_file).st_mtime) < 300 and \
-                    client_info.get('_html_ref') == exist.get('_html_ref') and \
-                    client_info.get('ip') == exist.get('ip'):
-                logger.critical(f"{exist.get('ip')} performed {exist.get('_html_ref')} once again within 5 minutes")
-                return
-
-        current_time = f"Server Datetime: {now.strftime('%B %d, %Y %I:%M %p')}"
-        hashes = ''.join(['#' for _ in range(74)])
-        with open(client_file, 'w') as client:
-            client.write(f"{hashes}\n#\t\t\t\t{current_time}\n{hashes}\n")
-            yaml.dump(client_info, client, indent=4)
-
-        if not (status := client_info.get('_html_ref')):
-            status = 'An undefined login attempt'
-
-        if not all([env.gmail_user, env.gmail_pass, env.recipient]) or status not in ['LOGIN SUCCESS',
-                                                                                      'An undefined login attempt']:
-            return
-
-        if not (endpoint := ngrok.get_ngrok()):
-            endpoint = f"http://{':'.join(map(str, self.connection.getsockname()))}"
-
-        with open(client_file, 'r') as client:
-            attachment_info = client.read()
-
-        body = f"A connection with status `{status}` has been made to your file server hosted at " \
-               f"{endpoint}\n\nDetails of the client are below. If this was you, you may disregard this email. " \
-               f"Otherwise stop the server immediately and rotate your credentials before restarting." \
-               f"\n\n\n{attachment_info}\n\n"
-
-        if env.host_dir == env.home_dir:
-            base_path = PurePath(__file__)
-            body += f"\n\n\nLogs: {endpoint}/" \
-                    f"{(str(base_path).strip(base_path.name) + models.LogConfig.SERVER_LOG_FILE).strip(env.host_dir)}"
-
-        response = SendEmail(gmail_user=env.gmail_user, gmail_pass=env.gmail_pass).send_email(
-            recipient=env.recipient, subject=f"WARNING: {status} was detected. {current_time}", attachment=client_file,
-            body=body, sender='FileWare'
-        )
-        if response.ok:
-            logger.info('Email has been sent successfully.')
-        else:
-            logger.error(response.body)
+        with open(filename, "w") as f:
+            yaml.dump(data=client_info, stream=f)
 
     def disable_cache(self) -> None:
         """Headers to force no-cache and site-data to expire."""
@@ -204,6 +153,7 @@ class Authenticator(SimpleHTTPRequestHandler):
         self.headers.add_header(_name='Clear-Site-Data', _value='"cache", "cookies", "storage", "executionContexts"')
 
 
+# noinspection PyProtectedMember
 def _reset_auth() -> bool:
     """Tells if the authentication header has to be reset and cache to be cleared.
 
@@ -214,9 +164,9 @@ def _reset_auth() -> bool:
         bool:
         True if it is the first login attempt, or it has been more than 15 minutes since the first/previous expiry.
     """
-    if env.first_run:
-        env.first_run = False
+    if settings._first_run:
+        settings._first_run = False
         return True
-    elif time() - env.start_time > 900:
-        env.start_time = time()
+    elif time() - settings._start_time > 900:
+        settings._start_time = time()
         return True