+++ gen_docs.sh
@@ -4,7 +4,6 @@
 set -e
 rm -rf docs
 mkdir docs
-python3 -m pip install --no-cache --upgrade sphinx
 [ ! -d "doc_generator/_static" ] && mkdir doc_generator/_static  # creates a _static folder if unavailable
 cp README.md doc_generator && cd doc_generator && make clean html && mv _build/html/* ../docs && rm README.md
 touch ../docs/.nojekyll